allows the height input to take floating values

+++ Client/src/utils/mutations.js
@@ -27,7 +27,7 @@ export const ADD_USER = gql`
 
 
 export const ADD_BODY = gql`
-    mutation addBody($weight: Float!, $bodyFatPercentage: Float!, $height: Int!){
+    mutation addBody($weight: Float!, $bodyFatPercentage: Float!, $height: Float!){
     addBody(weight: $weight, bodyFatPercentage: $bodyFatPercentage , height: $height ){
         username
     }
