+++ server/schemas/typeDefs.js
@@ -85,7 +85,7 @@ type Mutation {
     createUser(username: String!, email: String!, password: String!): Auth
     login(email: String!, password: String!): Auth
     
-    addBody(weight:Float!, bodyFatPercentage: Float!, height: Int!): User
+    addBody(weight:Float!, bodyFatPercentage: Float!, height: Float!): User
     addLift(exercise: String!, sets: [SetInput!]!): User
 
     sendPreSignedUrl(filename: String!, contentType: String!): String
